Refactor worker to handle errors a bit more gracefully

+++ app/workers/screenshot_worker.rb
@@ -7,13 +7,22 @@ class ScreenshotWorker
     name += Digest::SHA1.hexdigest([Time.now, rand].join)
 
     file_location = Rails.root.join("tmp/images/#{name}.png").to_s
-    ss.for(url).by_selector(selector).save!(file_location)
-    ss.free
 
-    f = File.new(file_location)
-    SnipUploader.new.store!(f)
-    File.delete(f.path)
-    Rails.logger.info "#{name} uploaded to s3"
+    if ss.for(url).by_selector(selector)
+      begin
+        ss.save!(file_location)
+        ss.free
+
+        f = File.new(file_location)
+        SnipUploader.new.store!(f)
+        File.delete(f.path)
+        Rails.logger.info "#{name} uploaded to s3"
+      rescue StandardError => e
+        Rails.logger.error "#{url} #{selector}: #{e}"
+      end
+    else
+      Rails.logger.info "#{name}: no selector matches"
+    end
   end
 
 end
